refactor(UserForm): extract to seperate file
