+++ modules/settings/Members.js
@@ -2,7 +2,8 @@ import React, { Component, PropTypes } from 'react'
 import _ from 'lodash'
 import AV from 'leancloud-storage/live-query'
 
-const {UserLabel, UserForm} = require('../common')
+const {UserLabel} = require('../common')
+import UserForm from './UserForm'
 import translate from '../i18n/translate'
 
 class SettingMembers extends Component {
